+++ services/madoc-ts/src/frontend/admin/pages/global/list-users.tsx
@@ -1,5 +1,5 @@
-import React, { useEffect, useState } from 'react';
-import { Link, Navigate, useNavigate } from 'react-router-dom';
+import React, { useEffect, useMemo, useState } from 'react';
+import { Navigate, useNavigate } from 'react-router-dom';
 import { getBotType } from '../../../../automation/utils/get-bot-type';
 import { User } from '../../../../extensions/site-manager/types';
 import { TimeAgo } from '../../../shared/atoms/TimeAgo';
@@ -9,7 +9,6 @@ import { SimpleTable } from '../../../shared/layout/SimpleTable';
 import { SuccessMessage } from '../../../shared/callouts/SuccessMessage';
 import { WidePage } from '../../../shared/layout/WidePage';
 import { usePaginatedData } from '../../../shared/hooks/use-data';
-import { useLocationQuery } from '../../../shared/hooks/use-location-query';
 import { useUser } from '../../../shared/hooks/use-site';
 import { serverRendererFor } from '../../../shared/plugins/external/server-renderer-for';
 import { HrefLink } from '../../../shared/utility/href-link';
@@ -18,13 +17,103 @@ import { Pagination } from '../../molecules/Pagination';
 import { Pagination as _Pagination } from '../../../../types/schemas/_pagination';
 import { validateEmail } from '../../../../utility/validate-email';
 import { ErrorMessage } from '../../../shared/callouts/ErrorMessage';
+import { ColumnDef, flexRender, getCoreRowModel, useReactTable, getSortedRowModel } from '@tanstack/react-table';
+import { ItemFilter } from '../../../shared/components/ItemFilter';
+import { useLocationState } from '../../../shared/hooks/use-location-state';
+import { SortableTableHeader } from '../../../shared/layout/SortableTableHeader';
+import { Input } from '../../../shared/form/Input';
+
+export const columns: ColumnDef<User>[] = [
+  {
+    accessorKey: 'id',
+    header: 'ID',
+    cell: info => info.getValue(),
+  },
+  {
+    accessorKey: 'name',
+    header: 'Name',
+    enableSorting: true,
+    cell: info => <HrefLink href={`/global/users/${info.row.original.id}`}>{info.getValue() as string}</HrefLink>,
+  },
+  {
+    accessorKey: 'email',
+    enableSorting: true,
+    header: 'Email',
+  },
+  {
+    accessorKey: 'is_active',
+    header: 'Is active',
+    cell: info => (info.getValue() ? 'active' : 'inactive'),
+  },
+  {
+    accessorKey: 'created',
+    header: 'Created',
+    cell: info => <TimeAgo date={new Date(info.getValue() as string)} />,
+  },
+  {
+    accessorKey: 'modified',
+    header: 'Last modified',
+    enableSorting: true,
+    cell: info => (info.getValue() ? <TimeAgo date={new Date(info.getValue() as string)} /> : '-'),
+  },
+  {
+    accessorKey: 'role',
+    header: 'Global role',
+    cell: info => <Tag>{info.getValue() as string}</Tag>,
+  },
+  {
+    id: 'automated',
+    header: 'Automated',
+    cell: info => {
+      const user = info.row.original;
+      return user.automated ? <Tag>{getBotType(user.config?.bot?.type) || 'bot'}</Tag> : null;
+    },
+  },
+];
 
 export const ListUsers: React.FC = () => {
   const { data } = usePaginatedData<{ users: User[]; pagination: _Pagination }>(ListUsers);
-  const query = useLocationQuery();
+  const [
+    query,
+    setQuery,
+    {
+      onSortingChange,
+      sortingState
+    }
+  ] = useLocationState();
+
+
   const currentUser = useUser();
   const navigate = useNavigate();
   const [userDeleted, setUserDeleted] = useState(false);
+
+  const activeFilter = query.status;
+  const setActiveFilter = (value: string) => {
+    setQuery({ ...query, status: value });
+  };
+
+  const roleFilter = query.role || null;
+  const setRoleFilter = (value: string | null) => {
+    setQuery({ ...query, role: value || undefined });
+  };
+
+  const filteredUsers = useMemo(() => {
+    if (!data?.users) return [];
+
+    return data.users.filter(user => {
+      if (roleFilter && user.role !== roleFilter) return false;
+      return !(activeFilter && (user.is_active ? 'active' : 'inactive') !== activeFilter);
+    });
+  }, [data?.users, roleFilter, activeFilter]);
+
+  const table = useReactTable({
+    data: filteredUsers,
+    columns,
+    state: { sorting: sortingState },
+    onSortingChange,
+    getCoreRowModel: getCoreRowModel(),
+    getSortedRowModel: getSortedRowModel(),
+  });
 
   useEffect(() => {
     if (query.user_deleted) {
@@ -63,75 +152,155 @@ export const ListUsers: React.FC = () => {
           <Button as={HrefLink} href={`/global/users/create-bot`}>
             Create bot
           </Button>
+          <ItemFilter
+            type="radio"
+            label={roleFilter ?? 'Filter by role'}
+            closeOnChange
+            items={[
+              { label: 'Global admin', id: 'global_admin' },
+              { label: 'Site admin', id: 'site_admin' },
+              { label: 'Editor', id: 'editor' }, // @todo remove?
+              { label: 'Reviewer', id: 'reviewer' }, // @todo remove?
+              { label: 'Author', id: 'author' }, // @todo remove?
+              { label: 'Researcher', id: 'researcher' }, // @todo remove?
+              { label: 'Transcriber', id: 'Transcriber' }, // @todo lowercase.
+            ].map(role => ({
+              id: role.id,
+              label: role.label,
+              onChange: selected => {
+                setRoleFilter(selected ? role.id : null);
+              },
+            }))}
+            selected={roleFilter ? [roleFilter] : []}
+          />
+
+          <ItemFilter
+            type="radio"
+            label={activeFilter ?? 'Filter by status'}
+            closeOnChange
+            items={[
+              { id: 'active', label: 'Active' },
+              { id: 'inactive', label: 'Inactive' },
+            ].map(s => ({
+              id: s.id,
+              label: s.label,
+              onChange: selected => {
+                setActiveFilter(selected ? (s.id as any) : null);
+              },
+            }))}
+            selected={activeFilter ? [activeFilter] : []}
+          />
+
+          <div className="px-3 w-full max-w-lg ml-auto">
+            <Input
+              value={query.search || ""}
+              onChange={e => setQuery({ search: e.target.value })}
+              placeholder="Search users..."
+            />
+          </div>
+
+          {(roleFilter || activeFilter) && (
+            <Button
+              onClick={() => {
+                setQuery({
+                  role: undefined,
+                  status: undefined,
+                  sort: undefined,
+                  search: undefined,
+                })
+              }}
+            >
+              Reset filters
+            </Button>
+          )}
         </ButtonRow>
-        <Pagination
-          page={data ? data.pagination.page : 1}
-          totalPages={data ? data.pagination.totalPages : 1}
-          stale={!data}
-        />
+        {(data?.pagination.totalPages || 0) > 1 && (
+          <Pagination
+            page={data ? data.pagination.page : 1}
+            totalPages={data ? data.pagination.totalPages : 1}
+            stale={!data}
+          />
+        )}
         <SimpleTable.Table>
           <thead>
-            <SimpleTable.Row>
-              <SimpleTable.Header>
-                <strong>ID</strong>
-              </SimpleTable.Header>
-              <SimpleTable.Header>
-                <strong>Name</strong>
-              </SimpleTable.Header>
-              <SimpleTable.Header>
-                <strong>Email</strong>
-              </SimpleTable.Header>
-              <SimpleTable.Header>
-                <strong>Is active</strong>
-              </SimpleTable.Header>
-              <SimpleTable.Header>
-                <strong>Created</strong>
-              </SimpleTable.Header>
-              <SimpleTable.Header>
-                <strong>Last modified</strong>
-              </SimpleTable.Header>
-              <SimpleTable.Header>
-                <strong>Global role</strong>
-              </SimpleTable.Header>
-              <SimpleTable.Header>
-                <strong>Automated</strong>
-              </SimpleTable.Header>
-            </SimpleTable.Row>
-          </thead>
-          <tbody>
-            {data?.users.map(user => (
-              <SimpleTable.Row key={user.id} $interactive style={{ background: user.automated || validateEmail(user.email) ? '' : '#FDD6D6' }}>
-                <SimpleTable.Cell>{user.id}</SimpleTable.Cell>
-                <SimpleTable.Cell>
-                  <HrefLink href={`/global/users/${user.id}`}>{user.name}</HrefLink>
-                </SimpleTable.Cell>
-                <SimpleTable.Cell>{user.email}</SimpleTable.Cell>
-                <SimpleTable.Cell>{user.is_active ? 'active' : 'inactive'}</SimpleTable.Cell>
-                <SimpleTable.Cell>
-                  <TimeAgo date={new Date(user.created)} />
-                </SimpleTable.Cell>
-                <SimpleTable.Cell>{user.modified ? <TimeAgo date={new Date(user.modified)} /> : '-'}</SimpleTable.Cell>
-                <SimpleTable.Cell>
-                  <Tag>{user.role}</Tag>
-                </SimpleTable.Cell>
-                <SimpleTable.Cell>
-                  {user.automated ? <Tag>{getBotType(user.config?.bot?.type) || 'bot'}</Tag> : null}
-                </SimpleTable.Cell>
+            {table.getHeaderGroups().map(headerGroup => (
+              <SimpleTable.Row key={headerGroup.id}>
+                {headerGroup.headers.map(header => (
+                  <SimpleTable.Header
+                    data-interactive={header.column.getCanSort()}
+                    data-pinned={header.column.getIsSorted() !== false}
+                    key={header.id}
+                    onClick={header.column.getCanSort() ? header.column.getToggleSortingHandler() : undefined}
+                    style={{
+                      cursor: header.column.getCanSort() ? 'pointer' : 'default',
+                      userSelect: 'none',
+                      whiteSpace: 'nowrap',
+                    }}
+                  >
+                    <SortableTableHeader header={header}>
+                      {flexRender(header.column.columnDef.header, header.getContext())}
+                    </SortableTableHeader>
+                  </SimpleTable.Header>
+                ))}
               </SimpleTable.Row>
             ))}
+          </thead>
+
+          <tbody>
+            {table.getRowModel().rows.map(row => {
+              const user = row.original;
+              const isInvalid = !user.automated && !validateEmail(user.email);
+              return (
+                <SimpleTable.Row key={row.id} $interactive style={{ background: isInvalid ? '#FDD6D6' : undefined }}>
+                  {row.getVisibleCells().map(cell => (
+                    <SimpleTable.Cell key={cell.id}>
+                      {flexRender(cell.column.columnDef.cell, cell.getContext())}
+                    </SimpleTable.Cell>
+                  ))}
+                </SimpleTable.Row>
+              );
+            })}
           </tbody>
         </SimpleTable.Table>
-        <Pagination
-          page={data ? data.pagination.page : 1}
-          totalPages={data ? data.pagination.totalPages : 1}
-          stale={!data}
-        />
+        {(data?.pagination.totalPages || 0) > 1 && (
+          <Pagination
+            page={data ? data.pagination.page : 1}
+            totalPages={data ? data.pagination.totalPages : 1}
+            stale={!data}
+          />
+        )}
       </WidePage>
     </>
   );
 };
 
 serverRendererFor(ListUsers, {
-  getKey: (params, query) => [`system-all-users`, { page: query.page ? Number(query.page) : 1 }],
-  getData: (key, vars, api) => api.siteManager.listAllUsers(vars.page),
+  getKey: (params, query) => [
+    `system-all-users`,
+    {
+      page: query.page ? Number(query.page) : 1,
+      roles: query.roles,
+      role: query.role,
+      status: query.status,
+      automated: query.automated === 'true' ? true : query.automated === 'false' ? false : undefined,
+      sort: query.sort || 'id',
+      asc: query.asc === 'true',
+      search: query.search || '',
+    },
+  ],
+  getData: (key, vars, api) =>
+    api.siteManager.listAllUsers(
+      vars.page,
+      {
+        automated: vars.automated,
+        role: vars.role,
+        roles: vars.roles,
+        status: vars.status,
+        search: vars.search,
+      },
+      {
+        direction: vars.asc ? 'asc' : 'desc',
+        name: vars.sort || 'id',
+      }
+    ),
 });